Add image for 768 < width <= 1024px, and fix bug.
The second image of homer now displays for this range of widths.
The bug was using [min/max]-device-width instead of [min/max]-width, which
meant that resizing the browser didn't cause any change, because the
device width always remains the same.  All references to -device-width have
been changed to -width.