Reduced repetition in CSS file
Made the background image link to the same file in all cases.
Made the dimentions of the div the same whenever width <= 1280px.
Changed all other rules to only refer to background-position, as this is the
only thing that changes between rules.
Page seems to respond exactly the same with the modified CSS file.